Refactor TreeData children into multiple files
Move custom hooks from TreeData to its children.
Move custom like and adoption hooks from TreeAdoptionLike.js to queries.js
Use custom hooks in TreeDataEdit.js
Enable pressing esc or clicking outside the modal to close it.
Simplify creation of data tables in <TreeLocation>, <TreeMoreInfo> and <TreeHeader>.
Support passing a default options object to createUseQuery().
Don't export putData, as it's no longer used directly outside of queries.js.
Fix typo "editted" in TreeDataEdit.js.
Format YMD dates with -, not /, in TreeDataEdit.js.
Rename treedata_utilities.js to treeDataUtils.js.
