+++ client/src/pages/treedata/TreeAdoptionLike.js
@@ -7,59 +7,80 @@ import StarIcon from '@material-ui/icons/Star';
 import StarBorderIcon from '@material-ui/icons/StarBorder';
 import format from 'date-fns/format';
 import React, { useState } from 'react';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { getData, postData } from '../../api/queries';
+import {
+  useTreeAdoptionsQuery, useTreeLikesQuery, useTreeAdoptionsMutation,
+  useTreeHistoryMutation, useTreeLikesMutation,
+} from '../../api/queries';
 import StarCheckbox from '../../components/Checkbox/StarCheckbox/StarCheckbox';
 import AdoptionInfoIconButton from '../../components/IconButtons/AdoptionInfoIconButton/AdoptionInfoIconButton';
 import AdoptionIcon from '../../components/Icons/AdoptionIcon/AdoptionIcon';
 import TreeAdoptionDirections from './TreeAdoptionDirections';
 
-export const useLikesQuery = (obj) => useQuery(['treelikes', obj], getData, {
-  placeholderData: {
-    liked: false,
-    likedCount: 0,
-  },
-});
+function Liked({ handleChange, idTree, user }) {
+  const { data: { liked, likedCount } } = useTreeLikesQuery({ idTree, email: user.email });
 
-export const useAdoptionQuery = (obj) => useQuery(['treeadoptions', obj], getData, {
-  placeholderData: {
-    adopted: false,
-    adoptedCount: 0,
-  },
-});
+  return (
+    <Grid item>
+      <Grid container alignItems="center">
+        <Grid item>
+          <Tooltip title={<div style={{ fontSize: '1.125rem' }}>Like</div>} placement="top" arrow>
+            <StarCheckbox
+              name="liked"
+              edge="start"
+              checked={liked}
+              icon={<StarBorderIcon fontSize="large" />}
+              checkedIcon={<StarIcon fontSize="large" />}
+              onChange={handleChange}
+            />
+          </Tooltip>
+        </Grid>
+        <Box fontSize="1.125rem">{likedCount}</Box>
+      </Grid>
+    </Grid>
+  );
+}
 
-export const useLikesMutation = () => {
-  const queryClient = useQueryClient();
+function Adopted({
+  handleChange, user, idTree, adoptionDirections, showAdoptionDirections,
+}) {
+  const { data: { adopted, adoptedCount } } = useTreeAdoptionsQuery({ idTree, email: user.email });
 
-  return useMutation(postData, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('treelikes');
-      queryClient.invalidateQueries('treehistory');
-    },
-    onError: (err) => {
-      console.error(err);
-    },
-  });
-};
+  return (
+    <Grid item>
+      <Grid container alignItems="center">
+        <Grid item>
+          <Tooltip title={<div style={{ fontSize: '1.125rem' }}>Adopt</div>} placement="top" arrow>
+            <Checkbox
+              name="adopted"
+              checked={adopted}
+              icon={<AdoptionIcon fontSize="large" />}
+              checkedIcon={<AdoptionIcon fontSize="large" primary />}
+              onChange={handleChange}
+            />
+          </Tooltip>
+        </Grid>
+        <Grid item>
+          <Box fontSize="1.125rem">{adoptedCount}</Box>
+        </Grid>
+        <Grid item>
+          <AdoptionInfoIconButton
+            edge="end"
+            adoptionDirections={adoptionDirections}
+            onClick={() => showAdoptionDirections(!adoptionDirections)}
+          >
+            <InfoIcon fontSize="large" />
+          </AdoptionInfoIconButton>
+        </Grid>
+      </Grid>
+    </Grid>
+  );
+}
 
-export const useAdoptionMutation = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation(postData, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('treeadoptions');
-      queryClient.invalidateQueries('treehistory');
-    },
-    onError: (err) => {
-      console.error(err);
-    },
-  });
-};
-
-export default function AdoptLikeCheckboxes({ idTree, common, mutateHistory }) {
+export default function AdoptLikeCheckboxes({ idTree, common }) {
   const { user } = useAuth0();
-  const mutateTreeLikes = useLikesMutation();
-  const mutateTreeAdoption = useAdoptionMutation();
+  const mutateTreeLikes = useTreeLikesMutation();
+  const mutateTreeAdoptions = useTreeAdoptionsMutation();
+  const mutateHistory = useTreeHistoryMutation();
   const [adoptionDirections, showAdoptionDirections] = useState(false);
 
   const handleChange = (event) => {
@@ -75,17 +96,16 @@ export default function AdoptLikeCheckboxes({ idTree, common, mutateHistory }) {
       nickname: user.nickname,
       email: user.email,
       request: {
-        // [event.target.name]: event.target.checked,
         name: event.target.name,
         type: event.target.checked ? 'POST' : 'DELETE',
       },
     };
 
     if (event.target.name === 'adopted') {
-      mutateTreeAdoption.mutate(['treeadoptions', sendTreeUser]);
+      mutateTreeAdoptions.mutate(sendTreeUser);
       showAdoptionDirections(event.target.checked);
     } else {
-      mutateTreeLikes.mutate(['treelikes', sendTreeUser]);
+      mutateTreeLikes.mutate(sendTreeUser);
     }
 
     mutateHistory.mutate(sendTreeHistory);
@@ -95,79 +115,21 @@ export default function AdoptLikeCheckboxes({ idTree, common, mutateHistory }) {
 
   return (
     <>
-      <Grid container alignItems="center" justify="space-between">
-        <Liked user={user} idTree={idTree} handleChange={handleChange} />
-        <Adopted
-          handleChange={handleChange}
-          user={user}
+      <Grid container alignItems="center" justifyContent="space-between">
+        <Liked
           idTree={idTree}
+          user={user}
+          handleChange={handleChange}
+        />
+        <Adopted
+          idTree={idTree}
+          user={user}
           adoptionDirections={adoptionDirections}
           showAdoptionDirections={showAdoptionDirections}
+          handleChange={handleChange}
         />
       </Grid>
       {adoptionDirections && <TreeAdoptionDirections />}
     </>
-  );
-}
-
-function Liked({ handleChange, idTree, user }) {
-  const { data } = useLikesQuery({ idTree, email: user.email });
-  const { liked, likedCount } = data;
-
-  return (
-    <Grid item>
-      <Grid container alignItems="center">
-        <Grid item>
-          <Tooltip title={<div style={{ fontSize: '1.125rem' }}>Like</div>} placement="top" arrow>
-            <StarCheckbox
-              edge="start"
-              icon={<StarBorderIcon fontSize="large" />}
-              checkedIcon={<StarIcon fontSize="large" />}
-              checked={liked}
-              onChange={handleChange}
-              name="liked"
-            />
-          </Tooltip>
-        </Grid>
-        <Box fontSize="1.125rem">{likedCount}</Box>
-      </Grid>
-    </Grid>
-  );
-}
-
-function Adopted({
-  handleChange, user, idTree, adoptionDirections, showAdoptionDirections,
-}) {
-  const { data } = useAdoptionQuery({ idTree, email: user.email });
-  const { adopted, adoptedCount } = data;
-
-  return (
-    <Grid item>
-      <Grid container alignItems="center">
-        <Grid item>
-          <Tooltip title={<div style={{ fontSize: '1.125rem' }}>Adopt</div>} placement="top" arrow>
-            <Checkbox
-              icon={<AdoptionIcon fontSize="large" />}
-              checkedIcon={<AdoptionIcon fontSize="large" primary />}
-              checked={adopted}
-              onChange={handleChange}
-              name="adopted"
-            />
-          </Tooltip>
-        </Grid>
-        <Grid item>
-          <Box fontSize="1.125rem">{adoptedCount}</Box>
-        </Grid>
-        <Grid item>
-          <AdoptionInfoIconButton
-            adoptionDirections={adoptionDirections}
-            edge="end"
-            onClick={() => showAdoptionDirections(!adoptionDirections)}
-          >
-            <InfoIcon fontSize="large" />
-          </AdoptionInfoIconButton>
-        </Grid>
-      </Grid>
-    </Grid>
   );
 }
