+++ client/src/pages/treedata/TreeHeader.js
@@ -1,58 +1,31 @@
 import React from 'react';
 import { Button } from 'reactstrap';
 import format from 'date-fns/format';
+import DataTable from './DataTable';
 
-export default function TreeHeader({
-  common, scientific, genus,
-  datePlanted, edit, height, dbh,
-}) {
+const headerKeys = [
+  ['common', '', 'h3'],
+  ['scientific', '', 'h4'],
+  // For many trees, the scientific and genus names are the same, so suppress that.
+  ['genus', '', 'h4', (data) => data.genus !== data.scientific && data.genus],
+  ['datePlanted', 'Planted', 'h5',
+    (data) => data.datePlanted
+      && !data.common.includes('VACANT')
+      && format(new Date(data.datePlanted), 'MMMM d, yyyy')],
+  ['height', undefined, 'h5'],
+  ['dbh', 'DBH', 'h5'],
+];
+
+export default function TreeHeader(props) {
   return (
     <div className="flex-grid-three text-left">
-      {common && (
-        <div>
-          <h3>{common}</h3>
-        </div>
-      )}
-      {scientific && (
-        <div>
-          <h4>{scientific}</h4>
-        </div>
-      )}
-      {genus && (
-        <div>
-          <h4>{genus}</h4>
-        </div>
-      )}
-      {datePlanted && !common.includes('VACANT') && (
-        <div>
-          <h5>
-            Planted:
-            {' '}
-            {format(new Date(datePlanted), 'MMMM dd, yyyy')}
-          </h5>
-        </div>
-      )}
-      {height && (
-        <div>
-          <h5>
-            Height:
-            {' '}
-            {height}
-          </h5>
-        </div>
-      )}
-      {dbh && (
-        <div>
-          <h5>
-            DBH:
-            {' '}
-            {dbh}
-          </h5>
-        </div>
-      )}
+      <DataTable
+        data={props}
+        keys={headerKeys}
+      />
 
       <div className="treedata__edit-btn text-right">
-        <Button color="link" className="btn-sm" onClick={edit}>Edit</Button>
+        <Button color="link" className="btn-sm" onClick={props.edit}>Edit</Button>
       </div>
     </div>
   );
