+++ client/src/pages/treedata/TreeInfo.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import DataTable from './DataTable';
+
+const locationKeys = ['address', 'city', 'zip', 'country', 'neighborhood', ['lng', 'Long'], 'lat'];
+const moreInfoKeys = [['owner', 'Organization'], 'who', ['idReference', 'Reference #']];
+
+export const TreeLocation = (props) => (
+  <div className="flex-grid border-top">
+    <div className="treehistory-list text-left">
+      <h4 className="text-center">Location</h4>
+      <DataTable
+        data={props}
+        keys={locationKeys}
+      />
+    </div>
+  </div>
+);
+
+export const TreeMoreInfo = (props) => (
+  <div className="flex-grid border-top">
+    <div className="treehistory-list text-left">
+      <h4 className="text-center">More info</h4>
+      <DataTable
+        data={props}
+        keys={moreInfoKeys}
+      />
+      <div>
+        <a href="https://standards.opencouncildata.org/#/trees" name="opencouncildata.org trees">
+          Open Tree Standards
+        </a>
+      </div>
+    </div>
+  </div>
+);
