+++ client/src/pages/treedata/TreeRemoval.js
@@ -1,13 +1,10 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { Button, ButtonGroup } from 'reactstrap';
 import format from 'date-fns/format';
-import './TreeData.scss';
 import { useAuth0 } from '@auth0/auth0-react';
+import { useTreeDataMutation, useTreeHistoryMutation } from '../../api/queries';
 
-export default function TreeRemoval({
-  idTree, common, notes,
-  mutateTreeData, mutateHistory,
-}) {
+export default function TreeRemoval({ idTree, common, notes }) {
   const { user, isAuthenticated, loginWithRedirect } = useAuth0();
   const mountedRef = useRef(true);
   const commentRef = useRef('');
@@ -16,6 +13,8 @@ export default function TreeRemoval({
   const [reasonForRemoval, setReasonForRemoval] = useState(false);
   const [reason, setReason] = useState(false);
   const [message, setMessage] = useState('Are you sure you want to remove this tree?');
+  const mutateTreeData = useTreeDataMutation();
+  const mutateHistory = useTreeHistoryMutation();
 
   const handleReason = () => {
     if (!isAuthenticated) loginWithRedirect();
@@ -38,7 +37,6 @@ export default function TreeRemoval({
       const comment = (commentRef.current && commentRef.current.value)
         ? commentRef.current.value
         : reason;
-
       const sendTreeHistory = {
         idTree,
         date_visit: dateVisit,
@@ -57,17 +55,20 @@ export default function TreeRemoval({
         // TODO: SHOULD NOTES OF PREVIOUS TREE BE DELETED or concated
       };
       const newNote = `${common} was removed by ${user.nickname} ${today} - ${comment}`;
+
       sendTreeData.notes = (notes && notes !== newNote)
         ? `${newNote}, ${notes}`
         : newNote;
-
       setMessage(`Removing ${common}.`);
 
+      // TODO: use mutateAsync() here to wait for it to finish?
       mutateHistory.mutate(sendTreeHistory);
       mutateTreeData.mutate(sendTreeData);
+
       setReallyDelete(false);
       setShowDelete(false);
       setMessage('');
+
       return sendTreeData;
     } catch (err) {
       console.error('CATCH', functionName, 'err', err);
@@ -100,7 +101,6 @@ export default function TreeRemoval({
           <div><h4>Reason For Removal?</h4></div>
           <span>
             <ButtonGroup className="treeremoval__reason-btngrp">
-
               <Button
                 name="NoWater"
                 className="treeremoval__reason-btn btn-md"
@@ -122,8 +122,8 @@ export default function TreeRemoval({
               >
                 Insects/Disease
               </Button>
-
             </ButtonGroup>
+
             <input
               ref={commentRef}
               placeholder="Other"
@@ -132,7 +132,6 @@ export default function TreeRemoval({
               className="treeremoval__reason-input"
               onBlur={handleRemoveTree}
             />
-
           </span>
         </div>
       )}
@@ -162,7 +161,6 @@ export default function TreeRemoval({
           </Button>
         </span>
       )}
-
     </div>
   );
 }
